+++ lib/src/levels/_levels.dart
@@ -1,5 +1,6 @@
 export 'conversation_level.dart';
 export 'credits_menu.dart';
+export 'custom_menu_level.dart';
 export 'main_menu.dart';
 export 'pause_menu.dart';
 export 'quest_menu.dart';